[FIFO/util] Move fifo resource modelling into utils
